# Conflicts:
#	README.md

+++ backend/jobsearch/src/main/kotlin/com/jobsearch/service/VacancyService.kt
@@ -1,10 +1,15 @@
 package com.jobsearch.service
 
-import com.jobsearch.dto.VacancyDto
+import com.jobsearch.dto.VacancyRequestDTO
+import com.jobsearch.dto.VacancyResponseDTO
 import com.jobsearch.entity.Vacancy
+import com.jobsearch.exception.ForbiddenException
+import com.jobsearch.exception.NotFoundException
 import com.jobsearch.repository.VacancyRepository
 import jakarta.transaction.Transactional
 import org.springframework.stereotype.Service
+import java.util.*
+import kotlin.jvm.optionals.getOrElse
 
 @Service
 class VacancyService(
@@ -12,42 +17,54 @@ class VacancyService(
     val jobFamilyService: JobFamilyService,
     val userService: UserService
 ) {
-    fun createVacancy(vacancyDto: VacancyDto): VacancyDto {
-        val selectedJobFamily = jobFamilyService.findByJobFamilyId(vacancyDto.jobFamilyId!!)
-            .orElseThrow { NoSuchElementException("No vacancy found with id ${vacancyDto.jobFamilyId}") }
+    fun retrieveVacancy(vacancyId: Int): VacancyResponseDTO {
+        val vacancy = vacancyRepository.findById(vacancyId)
+            .orElseThrow { NotFoundException("No vacancy found with id $vacancyId") }
+        return mapToVacancyResponseDto(vacancy)
+    }
 
+    fun retrieveAllVacancy(): List<VacancyResponseDTO> {
+        return vacancyRepository.findAll().map {
+            mapToVacancyResponseDto(it)
+        }
+    }
+
+    fun retrieveVacancyByManager(): List<VacancyResponseDTO> {
+        val manager = userService.retrieveAuthenticatedUser()
+        return vacancyRepository.findByManager(manager).map {
+            mapToVacancyResponseDto(it)
+        }
+    }
+    @Transactional
+    fun findVacanciesByFilter(salary: Int?, jobFamilyId: Int?, yearsOfExperience: Int?): List<VacancyResponseDTO> {
+        val vacancies = vacancyRepository.findVacanciesByFilters(salary, jobFamilyId, yearsOfExperience)
+        return vacancies.map {
+            mapToVacancyResponseDto(it)
+        }
+    }
+
+    @Transactional
+    fun createVacancy(vacancyDto: VacancyRequestDTO): VacancyResponseDTO {
+        val selectedJobFamily = jobFamilyService.findByJobFamilyId(vacancyDto.jobFamilyId)
         val managerUser = userService.retrieveAuthenticatedUser()
-
         val vacancyEntity = vacancyDto.let {
             Vacancy(it.id, it.name, it.companyName, it.salaryExpectation, it.yearsOfExperience, it.description, selectedJobFamily, managerUser)
         }
+
         val newVacancy = vacancyRepository.save(vacancyEntity)
 
-        return mapToVacancyDto(newVacancy)
+        return mapToVacancyResponseDto(newVacancy)
     }
 
-    fun retrieveVacancy(vacancyId: Int): VacancyDto {
+    @Transactional
+    fun updateVacancy(vacancyId: Int, vacancyDto: VacancyRequestDTO): VacancyResponseDTO {
+
         val vacancy = vacancyRepository.findById(vacancyId)
-            .orElseThrow { NoSuchElementException("No vacancy found with id $vacancyId") }
+            .orElseThrow { NotFoundException("No vacancy found with id $vacancyId") }
+        val manager = userService.retrieveAuthenticatedUser()
+        if (vacancy.manager.email != manager.email ) throw ForbiddenException("You are not allowed to edit this vacancy.")
 
-        return mapToVacancyDto(vacancy)
-    }
-
-    fun retrieveAllVacancy(): List<VacancyDto> {
-        val persons = vacancyRepository.findAll()
-
-        return persons.map {
-            mapToVacancyDto(it)
-        }
-    }
-
-    fun updateVacancy(vacancyId: Int, vacancyDto: VacancyDto): VacancyDto {
-        val vacancy = vacancyRepository.findById(vacancyId)
-            .orElseThrow { NoSuchElementException("No vacancy found with id $vacancyId") }
-
-        val selectedJobFamily = jobFamilyService.findByJobFamilyId(vacancyDto.jobFamilyId!!)
-            .orElseThrow { NoSuchElementException("No vacancy found with id ${vacancyDto.jobFamilyId}") }
-
+        val selectedJobFamily = jobFamilyService.findByJobFamilyId(vacancyDto.jobFamilyId)
         vacancy.name = vacancyDto.name
         vacancy.companyName = vacancyDto.companyName
         vacancy.salaryExpectation = vacancyDto.salaryExpectation
@@ -55,35 +72,35 @@ class VacancyService(
         vacancy.jobFamily = selectedJobFamily
 
         val updatedVacancy = vacancyRepository.save(vacancy)
-        return mapToVacancyDto(updatedVacancy)
-    }
 
-    fun deleteVacancy(vacancyId: Int): String {
-        val vacancy = vacancyRepository.findById(vacancyId)
-            .orElseThrow { NoSuchElementException("No vacancy found with id $vacancyId") }
-
-        vacancyRepository.delete(vacancy)
-        return "Vacancy deleted successfully"
+        return mapToVacancyResponseDto(updatedVacancy)
     }
 
     @Transactional
-    fun findVacanciesByFilter(salary: Int?, jobFamilyId: Int?, yearsOfExperience: Int?): List<VacancyDto> {
-        val vacancies = vacancyRepository.findVacanciesByFilters(salary, jobFamilyId, yearsOfExperience)
-        return vacancies.map {
-            mapToVacancyDto(it)
-        }
-    }
+    fun deleteVacancy(vacancyId: Int) {
+        val vacancy = vacancyRepository.findById(vacancyId)
+            .getOrElse { return }
+        val manager = userService.retrieveAuthenticatedUser()
+        if (vacancy.manager != manager ) throw ForbiddenException("You are not allowed to erase this vacancy.")
+        vacancyRepository.delete(vacancy)
+}
 
-    fun mapToVacancyDto(vacancy: Vacancy): VacancyDto {
+    /**
+     * Maps a Vacancy object to a VacancyResponseDTO object.
+     *
+     * @param vacancy the Vacancy object to map
+     * @return the mapped VacancyResponseDTO object
+     */
+    fun mapToVacancyResponseDto(vacancy: Vacancy): VacancyResponseDTO {
         return vacancy.let {
-            VacancyDto(
-                it.id,
+            VacancyResponseDTO(
+                it.id!!,
                 it.name,
                 it.companyName,
                 it.salaryExpectation,
                 it.yearsOfExperience,
                 it.description,
-                it.jobFamily.id,
+                it.jobFamily.id!!,
                 it.jobFamily.name,
                 it.manager.id
             )
